+++ src/controllers/lightbox/LightboxVideoCardController.js
@@ -1,8 +1,5 @@
 import VideoCardController from '../VideoCardController.js';
 import LightboxVideoCardView from '../../views/lightbox/LightboxVideoCardView.js';
-import DisplayAdVideoCardController from '../../mixins/DisplayAdVideoCardController.js';
-import ModalBallotResultsVideoCardController
-    from '../../mixins/ModalBallotResultsVideoCardController.js';
 import ModalShareVideoCardController from '../../mixins/ModalShareVideoCardController.js';
 
 export default class LightboxVideoCardController extends VideoCardController {
@@ -11,13 +8,9 @@ export default class LightboxVideoCardController extends VideoCardController {
 
         this.view = this.addView(new LightboxVideoCardView());
 
-        this.initDisplayAd();
-        this.initBallotResults();
         this.initShare();
     }
 }
 LightboxVideoCardController.mixin( // jshint ignore:line
-    DisplayAdVideoCardController,
-    ModalBallotResultsVideoCardController,
     ModalShareVideoCardController
 );