+++ src/solo-ads.js
@@ -1,5 +1,0 @@
-import './main.js';
-import Runner from '../lib/Runner.js';
-import SoloAdsApplicationController from './controllers/solo-ads/SoloAdsApplicationController.js';
-
-Runner.run(() => new SoloAdsApplicationController(document.body));