+++ src/swipe.js
@@ -1,6 +1,0 @@
-import './main.js';
-
-import Runner from '../lib/Runner.js';
-import SwipeApplicationController from './controllers/swipe/SwipeApplicationController.js';
-
-Runner.run(() => new SwipeApplicationController(document.body));