+++ src/views/full-np/FullNPInstagramImageCardView.js
@@ -1,0 +1,9 @@
+import InstagramCardView from '../InstagramCardView.js';
+
+export default class FullNPInstagramImageCardView extends InstagramCardView {
+    constructor() {
+        super(...arguments);
+
+        this.template = require('./FullNPInstagramImageCardView.html');
+    }
+}