+++ src/views/mobile/MobileTextCardView.js
@@ -1,9 +1,0 @@
-import CardView from '../CardView.js';
-
-export default class MobileTextCardView extends CardView {
-    constructor() {
-        super(...arguments);
-
-        this.template = require('./MobileTextCardView.html');
-    }
-}