+++ test/unit/spec/light/player.controller.light.ut.js
@@ -3,13 +3,9 @@ import LightPlayerController from '../../../../src/controllers/light/LightPlayer
 import PlayerController from '../../../../src/controllers/PlayerController.js';
 import LightPlayerView from '../../../../src/views/light/LightPlayerView.js';
 import ThumbnailNavigatorPlayerController from '../../../../src/mixins/ThumbnailNavigatorPlayerController.js';
-import LightArticleCardController from '../../../../src/controllers/light/LightArticleCardController.js';
-import LightTextCardController from '../../../../src/controllers/light/LightTextCardController.js';
 import LightImageCardController from '../../../../src/controllers/light/LightImageCardController.js';
 import LightVideoCardController from '../../../../src/controllers/light/LightVideoCardController.js';
-import LightboxPlaylistRecapCardController from '../../../../src/controllers/lightbox-playlist/LightboxPlaylistRecapCardController.js';
-import LightPrerollCardController from '../../../../src/controllers/light/LightPrerollCardController.js';
-import DisplayAdCardController from '../../../../src/controllers/DisplayAdCardController.js';
+import LightboxRecapCardController from '../../../../src/controllers/lightbox/LightboxRecapCardController.js';
 import LightInstagramImageCardController from '../../../../src/controllers/light/LightInstagramImageCardController.js';
 import LightInstagramVideoCardController from '../../../../src/controllers/light/LightInstagramVideoCardController.js';
 import dispatcher from '../../../../src/services/dispatcher.js';
@@ -56,18 +52,6 @@ describe('LightPlayerController', function() {
         });
 
         describe('CardControllers', function() {
-            describe('.article', function() {
-                it('should be LightArticleCardController', function() {
-                    expect(LightPlayerCtrl.CardControllers.article).toBe(LightArticleCardController);
-                });
-            });
-
-            describe('.text', function() {
-                it('should be LightTextCardController', function() {
-                    expect(LightPlayerCtrl.CardControllers.text).toBe(LightTextCardController);
-                });
-            });
-
             describe('.image', function() {
                 it('should be LightImageCardController', function() {
                     expect(LightPlayerCtrl.CardControllers.image).toBe(LightImageCardController);
@@ -81,20 +65,8 @@ describe('LightPlayerController', function() {
             });
 
             describe('.recap', function() {
-                it('should be LightboxPlaylistRecapCardController', function() {
-                    expect(LightPlayerCtrl.CardControllers.recap).toBe(LightboxPlaylistRecapCardController);
-                });
-            });
-
-            describe('.preroll', function() {
-                it('should be LightPrerollCardController', function() {
-                    expect(LightPlayerCtrl.CardControllers.preroll).toBe(LightPrerollCardController);
-                });
-            });
-
-            describe('.displayAd', function() {
-                it('should be DisplayAdCardController', function() {
-                    expect(LightPlayerCtrl.CardControllers.displayAd).toBe(DisplayAdCardController);
+                it('should be LightboxRecapCardController', function() {
+                    expect(LightPlayerCtrl.CardControllers.recap).toBe(LightboxRecapCardController);
                 });
             });