+++ test/unit/spec/lightbox/video_card.controller.lightbox.ut.js
@@ -2,8 +2,6 @@ import LightboxVideoCardController from '../../../../src/controllers/lightbox/Li
 import VideoCardController from '../../../../src/controllers/VideoCardController.js';
 import LightboxVideoCardView from '../../../../src/views/lightbox/LightboxVideoCardView.js';
 import {EventEmitter} from 'events';
-import DisplayAdVideoCardController from '../../../../src/mixins/DisplayAdVideoCardController.js';
-import ModalBallotResultsVideoCardController from '../../../../src/mixins/ModalBallotResultsVideoCardController.js';
 import ModalShareVideoCardController from '../../../../src/mixins/ModalShareVideoCardController.js';
 
 describe('LightboxVideoCardController', function() {
@@ -17,8 +15,6 @@ describe('LightboxVideoCardController', function() {
         card.thumbs = {};
         card.getSrc = jasmine.createSpy('card.getSrc()');
         spyOn(LightboxVideoCardController.prototype, 'addView').and.callThrough();
-        spyOn(LightboxVideoCardController.prototype, 'initDisplayAd').and.callThrough();
-        spyOn(LightboxVideoCardController.prototype, 'initBallotResults').and.callThrough();
         spyOn(LightboxVideoCardController.prototype, 'initShare').and.callThrough();
 
         LightboxVideoCardCtrl = new LightboxVideoCardController(card);
@@ -26,16 +22,6 @@ describe('LightboxVideoCardController', function() {
 
     it('should exist', function() {
         expect(LightboxVideoCardCtrl).toEqual(jasmine.any(VideoCardController));
-    });
-
-    it('should mixin the DisplayAdVideoCardController', function() {
-        expect(LightboxVideoCardController.mixins).toContain(DisplayAdVideoCardController);
-        expect(LightboxVideoCardCtrl.initDisplayAd).toHaveBeenCalled();
-    });
-
-    it('should mixin the ModalBallotResultsVideoCardController', function() {
-        expect(LightboxVideoCardController.mixins).toContain(ModalBallotResultsVideoCardController);
-        expect(LightboxVideoCardCtrl.initBallotResults).toHaveBeenCalled();
     });
 
     it('should mixin the ModalShareVideoCardController', function() {