+++ test/unit/spec/mobile/video_card.controller.mobile.ut.js
@@ -1,14 +1,9 @@
 import MobileVideoCardController from '../../../../src/controllers/mobile/MobileVideoCardController.js';
 import VideoCardController from '../../../../src/controllers/VideoCardController.js';
 import MobileVideoCardView from '../../../../src/views/mobile/MobileVideoCardView.js';
-import {EventEmitter} from 'events';
 import View from '../../../../lib/core/View.js';
 import VideoCard from '../../../../src/models/VideoCard.js';
 import playerFactory from '../../../../src/services/player_factory.js';
-import Runner from '../../../../lib/Runner.js';
-import DisplayAd from '../../../../src/models/DisplayAd.js';
-import DisplayAdController from '../../../../src/controllers/DisplayAdController.js';
-import InlineBallotResultsVideoCardController from '../../../../src/mixins/InlineBallotResultsVideoCardController.js';
 import ModalShareVideoCardController from '../../../../src/mixins/ModalShareVideoCardController.js';
 
 describe('MobileVideoCardController', function() {
@@ -90,7 +85,6 @@ describe('MobileVideoCardController', function() {
         spyOn(playerFactory, 'playerForCard').and.returnValue(player);
 
         spyOn(MobileVideoCardController.prototype, 'addView').and.callThrough();
-        spyOn(MobileVideoCardController.prototype, 'initBallotResults').and.callThrough();
         spyOn(MobileVideoCardController.prototype, 'initShare').and.callThrough();
 
         MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
@@ -98,11 +92,6 @@ describe('MobileVideoCardController', function() {
 
     it('should be a VideoCardController', function() {
         expect(MobileVideoCardCtrl).toEqual(jasmine.any(VideoCardController));
-    });
-
-    it('should mixin the InlineBallotResultsVideoCardController', function() {
-        expect(MobileVideoCardController.mixins).toContain(InlineBallotResultsVideoCardController);
-        expect(MobileVideoCardCtrl.initBallotResults).toHaveBeenCalled();
     });
 
     it('should mixin the ModalShareVideoCardController', function() {
@@ -115,249 +104,6 @@ describe('MobileVideoCardController', function() {
             it('should be a VideoCardView', function() {
                 expect(MobileVideoCardCtrl.view).toEqual(jasmine.any(MobileVideoCardView));
                 expect(MobileVideoCardCtrl.addView).toHaveBeenCalledWith(MobileVideoCardCtrl.view);
-            });
-        });
-
-        describe('DisplayAdCtrl', function() {
-            it('should not exist', function() {
-                expect('DisplayAdCtrl' in MobileVideoCardCtrl).toBe(false);
-            });
-
-            describe('if there is a displayAd', function() {
-                beforeEach(function() {
-                    card.modules.displayAd = {};
-                });
-
-                describe('if the placement is the default', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd.isDefault = true;
-
-                        MobileVideoCardCtrl = new MobileVideoCardController(card);
-                    });
-
-                    it('should not exist', function() {
-                        expect('DisplayAdCtrl' in MobileVideoCardCtrl).toBe(false);
-                    });
-                });
-
-                describe('if the placement is not the default', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd.isDefault = false;
-
-                        MobileVideoCardCtrl = new MobileVideoCardController(card);
-                    });
-
-                    it('should be a DisplayAdController', function() {
-                        expect(MobileVideoCardCtrl.DisplayAdCtrl).toEqual(jasmine.any(DisplayAdController));
-                    });
-                });
-            });
-        });
-    });
-
-    describe('events:', function() {
-        describe('DisplayAdCtrl', function() {
-            beforeEach(function() {
-                card.modules.displayAd = { isDefault: false };
-                MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-            });
-
-            describe('activate', function() {
-                beforeEach(function() {
-                    Runner.run(() => MobileVideoCardCtrl.view.create());
-                    spyOn(MobileVideoCardCtrl.view.playerOutlet, 'hide');
-                    spyOn(MobileVideoCardCtrl.view.replayContainer, 'show');
-
-                    MobileVideoCardCtrl.DisplayAdCtrl.emit('activate');
-                });
-
-                it('should hide the playerOutlet', function() {
-                    expect(MobileVideoCardCtrl.view.playerOutlet.hide).toHaveBeenCalled();
-                });
-
-                it('should show the replayContainer', function() {
-                    expect(MobileVideoCardCtrl.view.replayContainer.show).toHaveBeenCalled();
-                });
-            });
-
-            describe('deactivate', function() {
-                beforeEach(function() {
-                    Runner.run(() => MobileVideoCardCtrl.view.create());
-                    spyOn(MobileVideoCardCtrl.view.playerOutlet, 'show');
-                    spyOn(MobileVideoCardCtrl.view.replayContainer, 'hide');
-
-                    MobileVideoCardCtrl.DisplayAdCtrl.emit('deactivate');
-                });
-
-                it('should show the playerOutlet', function() {
-                    expect(MobileVideoCardCtrl.view.playerOutlet.show).toHaveBeenCalled();
-                });
-
-                it('should hide the replayContainer', function() {
-                    expect(MobileVideoCardCtrl.view.replayContainer.hide).toHaveBeenCalled();
-                });
-            });
-        });
-
-        describe('player', function() {
-            describe('play', function() {
-                beforeEach(function() {
-                    Runner.run(() => player.emit('play'));
-                });
-
-                it('should do nothing', function() {});
-
-                describe('if there is a DisplayAdCtrl', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd = { isDefault: false };
-                        MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                        spyOn(MobileVideoCardCtrl.DisplayAdCtrl, 'deactivate');
-
-                        Runner.run(() => player.emit('play'));
-                    });
-
-                    it('should deactivate the DisplayAdCtrl', function() {
-                        expect(MobileVideoCardCtrl.DisplayAdCtrl.deactivate).toHaveBeenCalled();
-                    });
-                });
-            });
-
-            describe('ended', function() {
-                describe('if the displayAd module is present', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd = { isDefault: false };
-
-                        MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                        spyOn(MobileVideoCardCtrl.DisplayAdCtrl, 'activate');
-
-                        Runner.run(() => player.emit('ended'));
-                    });
-
-                    it('should activate the DisplayAdCtrl', function() {
-                        expect(MobileVideoCardCtrl.DisplayAdCtrl.activate).toHaveBeenCalled();
-                    });
-                });
-            });
-        });
-    });
-
-    describe('methods:', function() {
-        describe('replay()', function() {
-            beforeEach(function() {
-                spyOn(VideoCardController.prototype, 'replay');
-
-                MobileVideoCardCtrl.replay();
-            });
-
-            it('should call super()', function() {
-                expect(VideoCardController.prototype.replay).toHaveBeenCalled();
-            });
-
-            describe('if there is a DisplayAdCtrl', function() {
-                beforeEach(function() {
-                    card.modules.displayAd = { isDefault: false };
-                    MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                    spyOn(MobileVideoCardCtrl.DisplayAdCtrl, 'deactivate');
-
-                    MobileVideoCardCtrl.replay();
-                });
-
-                it('should deactivate the DisplayAdCtrl', function() {
-                    expect(MobileVideoCardCtrl.DisplayAdCtrl.deactivate).toHaveBeenCalled();
-                });
-
-                it('should call super()', function() {
-                    expect(VideoCardController.prototype.replay).toHaveBeenCalled();
-                });
-            });
-        });
-
-        describe('canAutoadvance()', function() {
-            beforeEach(function() {
-                spyOn(VideoCardController.prototype, 'canAutoadvance').and.returnValue(true);
-            });
-
-            describe('if there is no displayAd module', function() {
-                beforeEach(function() {
-                    delete MobileVideoCardCtrl.DisplayAdCtrl;
-                    delete card.modules.displayAd;
-
-                    MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                });
-
-                it('should be true', function() {
-                    expect(MobileVideoCardCtrl.canAutoadvance()).toBe(true);
-                });
-
-                describe('if the parent implementation returns false', function() {
-                    beforeEach(function() {
-                        VideoCardController.prototype.canAutoadvance.and.returnValue(false);
-                    });
-
-                    it('should be false', function() {
-                        expect(MobileVideoCardCtrl.canAutoadvance()).toBe(false);
-                    });
-                });
-            });
-
-            describe('if there is a displayAd module', function() {
-                beforeEach(function() {
-                    MobileVideoCardCtrl.DisplayAdCtrl = new EventEmitter();
-                    card.modules.displayAd = new DisplayAd(card, experience);
-
-                    MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                });
-
-                describe('if the displayAd is the default placement', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd.isDefault = true;
-                    });
-
-                    it('should return true', function() {
-                        expect(MobileVideoCardCtrl.canAutoadvance()).toBe(true);
-                    });
-                });
-
-                describe('if the displayAd is not the default placement', function() {
-                    beforeEach(function() {
-                        card.modules.displayAd.isDefault = false;
-                    });
-
-                    it('should be false', function() {
-                        expect(MobileVideoCardCtrl.canAutoadvance()).toBe(false);
-                    });
-                });
-            });
-        });
-
-        describe('render()', function() {
-            beforeEach(function() {
-                spyOn(VideoCardController.prototype, 'render');
-            });
-
-            describe('if there is no displayAd', function() {
-                beforeEach(function() {
-                    Runner.run(() => MobileVideoCardCtrl.render());
-                });
-
-                it('should call super()', function() {
-                    expect(VideoCardController.prototype.render).toHaveBeenCalled();
-                });
-            });
-
-            describe('if there is no displayAd', function() {
-                beforeEach(function() {
-                    card.modules.displayAd = { isDefault: false };
-                    MobileVideoCardCtrl = new MobileVideoCardController(card, parentView);
-                    Runner.run(() => MobileVideoCardCtrl.view.create());
-                    spyOn(MobileVideoCardCtrl.DisplayAdCtrl, 'renderInto');
-
-                    Runner.run(() => MobileVideoCardCtrl.render());
-                });
-
-                it('should call super()', function() {
-                    expect(MobileVideoCardCtrl.DisplayAdCtrl.renderInto).toHaveBeenCalledWith(MobileVideoCardCtrl.view.displayAdOutlet);
-                });
             });
         });
     });