+++ test/unit/spec/thumbnail_navigator_player_controller.mixin.ut.js
@@ -5,8 +5,6 @@ import PlayerView from '../../../src/views/PlayerView.js';
 import View from '../../../lib/core/View.js';
 import Runner from '../../../lib/Runner.js';
 import LinksListView from '../../../src/views/LinksListView.js';
-import PrerollCardController from '../../../src/controllers/PrerollCardController.js';
-import PrerollCard from '../../../src/models/PrerollCard.js';
 import dispatcher from '../../../src/services/dispatcher.js';
 class DeckView extends View {
     show() {}
@@ -29,7 +27,6 @@ describe('ThumbnailNavigatorPlayerController', function() {
         view = new PlayerView();
 
         Ctrl = new MyPlayerController(new View(document.createElement('body')));
-        Ctrl.CardControllers.preroll = PrerollCardController;
         Ctrl.view = view;
     });
 
@@ -57,11 +54,9 @@ describe('ThumbnailNavigatorPlayerController', function() {
                             view.pagerOutlet = new View();
                             view.links = new LinksListView();
                             spyOn(Ctrl.ThumbnailNavigatorViewCtrl, 'renderInto');
-                            spyOn(PrerollCardController.prototype, 'renderInto');
                             Ctrl.minireel.adConfig = {
                                 video: {}
                             };
-                            Ctrl.minireel.prerollCard = new PrerollCard({ data: {}, collateral: {}, params: {} }, experience, profile, Ctrl.minireel);
                             Ctrl.minireel.length = 5;
                             Ctrl.minireel.campaign = {};
 
@@ -94,7 +89,7 @@ describe('ThumbnailNavigatorPlayerController', function() {
                             view.links = new LinksListView();
                         });
 
-                        ['recap', 'text', 'displayAd', 'article', 'instagramImage', 'instagramVideo'].forEach(type => {
+                        ['recap', 'instagramImage', 'instagramVideo'].forEach(type => {
                             describe(`if the currentCard is a ${type} card`, function() {
                                 beforeEach(function() {
                                     Ctrl.minireel.currentCard = { type };
@@ -148,7 +143,7 @@ describe('ThumbnailNavigatorPlayerController', function() {
                             });
                         });
 
-                        ['video', 'text'].forEach(type => {
+                        ['video'].forEach(type => {
                             describe(`if the currentCard is a ${type} card`, function() {
                                 beforeEach(function() {
                                     Ctrl.minireel.currentCard = { type };