+++ test/unit/spec/video_card.controller.ut.js
@@ -8,7 +8,6 @@ import VideoCardView from '../../../src/views/VideoCardView.js';
 import playerFactory from '../../../src/services/player_factory.js';
 import dispatcher from '../../../src/services/dispatcher.js';
 import PostVideoCardController from '../../../src/mixins/PostVideoCardController.js';
-import BallotVideoCardController from '../../../src/mixins/BallotVideoCardController.js';
 import SponsoredCardController from '../../../src/mixins/SponsoredCardController.js';
 
 describe('VideoCardController', function() {
@@ -99,7 +98,6 @@ describe('VideoCardController', function() {
         spyOn(dispatcher, 'addSource');
 
         spyOn(VideoCardController.prototype, 'initPost').and.callThrough();
-        spyOn(VideoCardController.prototype, 'initBallot').and.callThrough();
 
         VideoCardCtrl = new VideoCardController(card, parentView);
         VideoCardCtrl.view = new VideoCardView();
@@ -112,11 +110,6 @@ describe('VideoCardController', function() {
     it('should mixin the PostVideoCardController', function() {
         expect(VideoCardController.mixins).toContain(PostVideoCardController);
         expect(VideoCardCtrl.initPost).toHaveBeenCalled();
-    });
-
-    it('should mixin the BallotVideoCardController', function() {
-        expect(VideoCardController.mixins).toContain(BallotVideoCardController);
-        expect(VideoCardCtrl.initBallot).toHaveBeenCalled();
     });
 
     it('should mixin the SponsoredCardController', function() {